# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Managing tool for album records",
        "title": "Album Store API",
        "contact": {
            "name": "API Support",
            "email": "arian.khanjani@gmail.com"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
        "/v1/albums": {
            "get": {
                "description": "Get details of all albums",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all albums",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Album"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new album with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new album record",
                "parameters": [
                    {
                        "description": "Create Album",
                        "name": "models.Album",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Album"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Album"
                        }
                    }
                }
            }
        },
        "/v1/albums/{id}": {
            "get": {
                "description": "Get details of an album",
                "produces": [
                    "application/json"
                ],
                "summary": "Get details of an album",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Album ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Album"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Album": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}