# models/album.go
package models

// Album represents data about a record album.
type Album struct {
	ID     string  `json:"id" gorm:"primary_key"`
	Title  string  `json:"title"`
	Artist string  `json:"artist"`
	Price  float64 `json:"price"`
}

// CreateAlbumInput represents the structure of creating an album record
type CreateAlbumInput struct {
	Title string `json:"title" binding:"required"`
	Artist string `json:"artist" binding:"required"`
	Price float64 `json:"price" binding:"required"`
}

// UpdateAlbumInput represents the structure of updating an album record
type UpdateAlbumInput struct {
	Title string `json:"title"`
	Artist string `json:"artist"`
	Price float64 `json:"price"`
}

// Albums slice to seed record album data.
var Albums = []Album{
	{ID: "1", Title: "Blue Train", Artist: "John Coltrane", Price: 56.99},
	{ID: "2", Title: "Jerry", Artist: "Gerry Mulligan", Price: 17.99},
	{ID: "3", Title: "Sarah Vaughan and Clifford Brown", Artist: "Sarah Vaughan", Price: 39.99},
}